fix: process video packets in order
This prevents a situation where we would permanently break a
VTDecompressionSession by feeding it out-of-order packets, causing the
attachment to freeze.

The solution is not great, because it means moving the packet processing
onto the main thread. I want to move a bunch of this into the rust
portion of the client anyway, at which point that won't be a concern
anymore.