fix: discard packets when they would break the decoder
This fixes another instance of the decoder freezing, which would happen
when we drop a packet with hierarchical layer N and then feed the
decoder a packet depending on it (always hierarchical layer >N).

The fix is to discard frames from higher layers until we "recover" by
receiving a packet with a lower layer.